+++ changelog.md
@@ -4,6 +4,7 @@
 ### New/Changed Features
 
 * Support for simple defines (like `#define FOO 1`)
+* Support for translation of preprocessor constants in array sizes.
 * Basic unit tests were added
 * Most of cucumber tests was replaced with D-based tests.
 * Statements are translated in original 'C' order now.